合入PUCCH上报接口

+++ bbmain_ctrl/phy_ctrl/ul_ctrl/ul_ctr_proc/ulctr_main.c
@@ -12,7 +12,7 @@ int main(void)
 {
   uint16_t a = 16;
   uint16_t b = 16;
-  uint16_t c = 0;
+  uint32_t c = 0;
 
   printf("c = %d;\n",c);
   printf("___Hello World___;\n");
